Move SourceType to SearchParams
